+++ src/Api/Providers/News.php
@@ -3,6 +3,7 @@
 namespace seregazhuk\PinterestBot\Api\Providers;
 
 use seregazhuk\PinterestBot\Api\Traits\HasFeed;
+use seregazhuk\PinterestBot\Helpers\Pagination;
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
 
 class News extends Provider
@@ -20,7 +21,7 @@ class News extends Provider
      * @param int $limit
      * @return mixed
      */
-    public function all($limit = 0)
+    public function all($limit = Pagination::DEFAULT_LIMIT)
     {
         $data = ['allow_stale' => true];