+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 
 ## [5.1.1] - 2016-12-03
+### Added:
+ - To any pagination added *toArray* method to receive all pagination results as array. 
 
 ### Fixed:
  - Pins *activity* now returns an empty error for no results instead of null.